refactor
move TransformToTemplateModel to buildPage
