--- src/docfx/template/TemplateUtility.cs
@@ -1,84 +1,0 @@
-// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Web;
-using Newtonsoft.Json.Linq;
-
-namespace Microsoft.Docs.Build
-{
-    internal class TemplateUtility
-    {
-        public static JObject CreateMetadata(JObject rawMetadata)
-        {
-            var metadata = new JObject();
-
-            foreach (var (key, value) in rawMetadata)
-            {
-                if (!key.StartsWith("_"))
-                {
-                    metadata[key] = value;
-                }
-            }
-
-            metadata["is_dynamic_rendering"] = true;
-
-            return metadata;
-        }
-
-        public static string CreateHtmlMetaTags(JObject metadata, HashSet<string> htmlMetaHidden, Dictionary<string, string> htmlMetaNames)
-        {
-            var result = new StringBuilder();
-
-            foreach (var property in metadata.Properties().OrderBy(p => p.Name))
-            {
-                var key = property.Name;
-                var value = property.Value;
-                if (value is JObject || htmlMetaHidden.Contains(key))
-                {
-                    continue;
-                }
-
-                var content = "";
-                var name = htmlMetaNames.TryGetValue(key, out var diplayName) ? diplayName : key;
-
-                if (value is JArray arr)
-                {
-                    foreach (var v in value)
-                    {
-                        if (v is JValue)
-                        {
-                            result.AppendLine($"<meta name=\"{HttpUtility.HtmlEncode(name)}\" content=\"{HttpUtility.HtmlEncode(v)}\" />");
-                        }
-                    }
-                    continue;
-                }
-                else if (value.Type == JTokenType.Boolean)
-                {
-                    content = (bool)value ? "true" : "false";
-                }
-                else
-                {
-                    content = value.ToString();
-                }
-
-                result.AppendLine($"<meta name=\"{HttpUtility.HtmlEncode(name)}\" content=\"{HttpUtility.HtmlEncode(content)}\" />");
-            }
-
-            return result.ToString();
-        }
-
-        public static JObject RemoveUpdatedAtDateTime(JObject rawMetadata)
-        {
-            JToken gitContributorInformation;
-            if (rawMetadata.TryGetValue("_op_gitContributorInformation", out gitContributorInformation)
-                && ((JObject)gitContributorInformation).ContainsKey("updated_at_date_time"))
-            {
-                ((JObject)rawMetadata["_op_gitContributorInformation"]).Remove("updated_at_date_time");
-            }
-            return rawMetadata;
-        }
-    }
-}
